project/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==4.0.3
django-extensions==3.2.3
idna==3.6
MarkupSafe==2.1.5
Naked==0.1.32
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
sqlparse==0.4.4
urllib3==2.2.1
Werkzeug==3.0.1
